requirements.txt
torch
captum
numpy
pandas
matplotlib
seaborn
bioframe
scipy
tqdm
loguru
pysam
pyBigWig
pyyaml
